Removed XMPMeta interface and merged this with the real class.
