+++ src/commonMain/kotlin/com/ashampoo/xmp/impl/XMPRDFParser.kt
@@ -11,6 +11,7 @@ package com.ashampoo.xmp.impl
 import com.ashampoo.xmp.XMPConst
 import com.ashampoo.xmp.XMPError
 import com.ashampoo.xmp.XMPException
+import com.ashampoo.xmp.XMPMeta
 import com.ashampoo.xmp.XMPMetaFactory.schemaRegistry
 import com.ashampoo.xmp.options.ParseOptions
 import com.ashampoo.xmp.options.PropertyOptions
@@ -106,9 +107,9 @@ internal object XMPRDFParser {
      *
      */
     @kotlin.jvm.JvmStatic
-    fun parse(xmlRoot: Node, options: ParseOptions): XMPMetaImpl {
+    fun parse(xmlRoot: Node, options: ParseOptions): XMPMeta {
 
-        val xmp = XMPMetaImpl()
+        val xmp = XMPMeta()
 
         parseRdfRoot(xmp, xmlRoot, options)
 
@@ -121,7 +122,7 @@ internal object XMPRDFParser {
      * They simply return for success, failures will throw an exception.
      */
     @Suppress("ThrowsCount")
-    fun parseRdfRoot(xmp: XMPMetaImpl, rdfRdfNode: Node, options: ParseOptions) {
+    fun parseRdfRoot(xmp: XMPMeta, rdfRdfNode: Node, options: ParseOptions) {
 
         if (rdfRdfNode.nodeName != "rdf:RDF")
             throw XMPException("Root node should be of type rdf:RDF", XMPError.BADRDF)
@@ -158,7 +159,7 @@ internal object XMPRDFParser {
      * term.
      */
     private fun parseRdfNodeElement(
-        xmp: XMPMetaImpl,
+        xmp: XMPMeta,
         xmpParent: XMPNode,
         xmlNode: Element,
         isTopLevel: Boolean,
@@ -192,7 +193,7 @@ internal object XMPRDFParser {
      */
     @Suppress("ThrowsCount")
     private fun parseRdfNodeElementAttrs(
-        xmp: XMPMetaImpl,
+        xmp: XMPMeta,
         xmpParent: XMPNode,
         xmlNode: Element,
         isTopLevel: Boolean
@@ -258,7 +259,7 @@ internal object XMPRDFParser {
      *
      */
     private fun parseRdfPropertyElementList(
-        xmp: XMPMetaImpl,
+        xmp: XMPMeta,
         xmpParent: XMPNode,
         xmlParent: Node?,
         isTopLevel: Boolean,
@@ -334,7 +335,7 @@ internal object XMPRDFParser {
      *
      */
     private fun parseRdfPropertyElement(
-        xmp: XMPMetaImpl,
+        xmp: XMPMeta,
         xmpParent: XMPNode,
         xmlNode: Element,
         isTopLevel: Boolean,
@@ -454,7 +455,7 @@ internal object XMPRDFParser {
      * properties written with rdf:Description and rdf:value.
      */
     private fun parseRdfResourcePropertyElement(
-        xmp: XMPMetaImpl,
+        xmp: XMPMeta,
         xmpParent: XMPNode,
         xmlNode: Element,
         isTopLevel: Boolean,
@@ -562,7 +563,7 @@ internal object XMPRDFParser {
      * Add a leaf node with the text value and qualifiers for the attributes.
      */
     private fun parseRdfLiteralPropertyElement(
-        xmp: XMPMetaImpl,
+        xmp: XMPMeta,
         xmpParent: XMPNode,
         xmlNode: Element,
         isTopLevel: Boolean
@@ -615,7 +616,7 @@ internal object XMPRDFParser {
      * Then process the XML child nodes to get the struct fields.
      */
     private fun parseTypeResourcePropertyElement(
-        xmp: XMPMetaImpl,
+        xmp: XMPMeta,
         xmpParent: XMPNode,
         xmlNode: Element,
         isTopLevel: Boolean,
@@ -688,7 +689,7 @@ internal object XMPRDFParser {
      *  4. Otherwise this is a struct, the attributes other than xml:lang, rdf:ID, or rdf:nodeID are fields.
      */
     private fun parseEmptyPropertyElement(
-        xmp: XMPMetaImpl,
+        xmp: XMPMeta,
         xmpParent: XMPNode,
         xmlNode: Element,
         isTopLevel: Boolean
@@ -846,7 +847,7 @@ internal object XMPRDFParser {
     }
 
     private fun addChildNode(
-        xmp: XMPMetaImpl,
+        xmp: XMPMeta,
         xmpParent: XMPNode,
         xmlNode: Node,
         value: String?,
